+++ README.md
@@ -87,3 +87,5 @@
 - indent depth를 3이 넘지 않도록 구현한다.
 - else 예약어를 쓰지 않는다.
 - Java Enum을 적용하여 구현한다.
+
+### 예외 발생 상황
