+++ src/main/java/lotto/controller/LottoController.java
@@ -1,22 +1,24 @@
 package lotto.controller;
 
 import lotto.model.UserLotto;
-import lotto.service.LottoService;
+import lotto.service.PurchaseService;
 import lotto.service.ProfitService;
+import lotto.service.WinningService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 import java.util.List;
 
 public class LottoController {
-    LottoService lottoService = new LottoService();
-    ProfitService profitService = new ProfitService();
     UserLotto userLotto = new UserLotto();
+    PurchaseService purchaseService = new PurchaseService();
+    WinningService winningService = new WinningService();
+    ProfitService profitService = new ProfitService();
 
     public void run() {
         String purchaseAmount = InputView.readPurchaseAmount();
-        int lottoCount = lottoService.calculateLottoCount(purchaseAmount);
-        List<String> purchasedLottoNumber = lottoService.purchaseLotto(lottoCount);
+        int lottoCount = purchaseService.calculateLottoCount(purchaseAmount);
+        List<String> purchasedLottoNumber = purchaseService.purchaseLotto(lottoCount);
         OutputView.printPurchasedLotto(lottoCount, purchasedLottoNumber);
 
         String userNumber = InputView.readUserNumber();
@@ -25,7 +27,7 @@ public class LottoController {
         String bonusNumber = InputView.readBonusNumber();
         userLotto.setBonusNumber(bonusNumber);
 
-        lottoService.matchLotto(userLotto);
+        winningService.matchLotto(purchaseService.getPurchasedLotto(), userLotto);
         OutputView.printWinningResult();
 
         double profitRate = profitService.calculate(lottoCount);
