+++ src/test/java/lotto/model/LottoTest.java
@@ -1,7 +1,6 @@
 package lotto.model;
 
 import org.junit.jupiter.api.Test;
-
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
